+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add automatic Chromecast/Google Cast output discovery and playback.
 - Add automatic UPnP MediaRenderer output discovery and playback.
 
 ## 1.1 (2026-05-30)